+++ source/Domain/GeoJSON/FeatureCollection.ts
@@ -1,16 +1,46 @@
-import { all, isArrayOfType, isKeyOfType } from '@konfirm/guard';
+import {
+	all,
+	any,
+	type Guard,
+	isArrayOfType,
+	isKeyOfType,
+	isNULL,
+} from '@konfirm/guard';
 import { type GeoJSONObject, isGeoJSONObject } from './Concept/GeoJSONObject';
 import { type Feature, isFeature, isStrictFeature } from './Feature';
+import { type Geometry, isGeometry } from './Geometry';
 
-export type FeatureCollection = GeoJSONObject<{
-	type: 'FeatureCollection';
-	features: Array<Feature>;
-}>;
-export const isFeatureCollection = all<FeatureCollection>(
+export type FeatureCollection<G extends Geometry | null = Geometry | null> =
+	GeoJSONObject<{
+		type: 'FeatureCollection';
+		features: Array<Feature<G>>;
+	}>;
+
+const isFeatureCollectionObject = all<FeatureCollection>(
 	isGeoJSONObject('FeatureCollection'),
 	isKeyOfType('features', isArrayOfType(isFeature)),
 );
-export const isStrictFeatureCollection = all<FeatureCollection>(
+const isStrictFeatureCollectionObject = all<FeatureCollection>(
 	isGeoJSONObject('FeatureCollection'),
 	isKeyOfType('features', isArrayOfType(isStrictFeature)),
 );
+
+export function isFeatureCollection<G extends Geometry | null>(
+	value: unknown,
+	isG: Guard<G> = any(isNULL, isGeometry),
+): value is FeatureCollection<G> {
+	return (
+		isFeatureCollectionObject(value) &&
+		value.features.every((feature) => isFeature(feature, isG))
+	);
+}
+
+export function isStrictFeatureCollection<G extends Geometry | null>(
+	value: unknown,
+	isG: Guard<G> = any(isNULL, isGeometry),
+): value is FeatureCollection<G> {
+	return (
+		isStrictFeatureCollectionObject(value) &&
+		value.features.every((feature) => isStrictFeature(feature, isG))
+	);
+}